sender: refactor sender framework
