+++ service/cronjob/monitor/monitor.go
@@ -14,7 +14,7 @@ import (
 )
 
 var configFile = flag.String("f",
-	"./etc/monitor.yaml", "the config file")
+	"./etc/config.yaml", "the path of config file")
 
 func main() {
 	flag.Parse()
