+++ service/cronjob/sender/svc/types.go
@@ -1,4 +1,20 @@
-package logic
+/*
+ * Copyright © 2021 Zkbas Protocol
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package svc
 
 import (
 	"encoding/json"
@@ -7,33 +23,65 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/zeromicro/go-zero/core/logx"
 
+	"github.com/bnb-chain/zkbas-eth-rpc/_rpc"
+	zkbas "github.com/bnb-chain/zkbas-eth-rpc/zkbas/core/legend"
+	"github.com/bnb-chain/zkbas/common/model/block"
+	"github.com/bnb-chain/zkbas/common/model/blockForCommit"
+	"github.com/bnb-chain/zkbas/common/model/l1TxSender"
+	"github.com/bnb-chain/zkbas/common/model/proofSender"
 	"github.com/bnb-chain/zkbas/common/tree"
 	"github.com/bnb-chain/zkbas/common/util"
 )
 
+type (
+	Block               = block.Block
+	BlockForCommit      = blockForCommit.BlockForCommit
+	L1TxSenderModel     = l1TxSender.L1TxSenderModel
+	L1TxSender          = l1TxSender.L1TxSender
+	BlockModel          = block.BlockModel
+	BlockForCommitModel = blockForCommit.BlockForCommitModel
+
+	ProviderClient = _rpc.ProviderClient
+	AuthClient     = _rpc.AuthClient
+	Zkbas          = zkbas.Zkbas
+
+	ZkbasCommitBlockInfo   = zkbas.OldZkbasCommitBlockInfo
+	ZkbasVerifyBlockInfo   = zkbas.OldZkbasVerifyAndExecuteBlockInfo
+	StorageStoredBlockInfo = zkbas.StorageStoredBlockInfo
+
+	ProofSenderModel = proofSender.ProofSenderModel
+)
+
+type SenderParam struct {
+	Cli            *ProviderClient
+	AuthCli        *AuthClient
+	ZkbasInstance  *Zkbas
+	MaxWaitingTime int64
+	MaxBlocksCount int
+	GasPrice       *big.Int
+	GasLimit       uint64
+}
+
 func DefaultBlockHeader() StorageStoredBlockInfo {
 	var (
-		pendingOnchainOperationsHash [32]byte
+		pendingOnChainOperationsHash [32]byte
 		stateRoot                    [32]byte
 		commitment                   [32]byte
 	)
-	copy(pendingOnchainOperationsHash[:], common.FromHex(util.EmptyStringKeccak)[:])
+	copy(pendingOnChainOperationsHash[:], common.FromHex(util.EmptyStringKeccak)[:])
 	copy(stateRoot[:], tree.NilStateRoot[:])
 	copy(commitment[:], common.FromHex("0x0000000000000000000000000000000000000000000000000000000000000000")[:])
 	return StorageStoredBlockInfo{
 		BlockSize:                    0,
 		BlockNumber:                  0,
 		PriorityOperations:           0,
-		PendingOnchainOperationsHash: pendingOnchainOperationsHash,
+		PendingOnchainOperationsHash: pendingOnChainOperationsHash,
 		Timestamp:                    big.NewInt(0),
 		StateRoot:                    stateRoot,
 		Commitment:                   commitment,
 	}
 }
 
-/*
-	ConvertBlocksForCommitToCommitBlockInfos: helper function to convert blocks to commit block infos
-*/
 func ConvertBlocksForCommitToCommitBlockInfos(oBlocks []*BlockForCommit) (commitBlocks []ZkbasCommitBlockInfo, err error) {
 	for _, oBlock := range oBlocks {
 		var newStateRoot [32]byte
